argminmax::scalar

Trait ScalarArgMinMax

pub trait ScalarArgMinMax<ScalarDType: Copy + PartialOrd> {
    // Required methods
    fn argminmax(data: &[ScalarDType]) -> (usize, usize);
    fn argmin(data: &[ScalarDType]) -> usize;
    fn argmax(data: &[ScalarDType]) -> usize;
}

A trait providing the scalar implementation of the argminmax operations.

Required Methods

fn argminmax(data: &[ScalarDType]) -> (usize, usize)

Get the index of the minimum and maximum values in the slice.

Arguments

data - the slice of data.

Returns

A tuple of the index of the minimum and maximum values in the slice (min_index, max_index).

fn argmin(data: &[ScalarDType]) -> usize

Get the index of the minimum value in the slice.

Arguments

data - the slice of data.

Returns

The index of the minimum value in the slice.

fn argmax(data: &[ScalarDType]) -> usize

Get the index of the maximum value in the slice.

Arguments

data - the slice of data.

Returns

The index of the maximum value in the slice.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.