argminmax

Trait NaNArgMinMax

pub trait NaNArgMinMax {
    // Required methods
    fn nanargminmax(&self) -> (usize, usize);
    fn nanargmin(&self) -> usize;
    fn nanargmax(&self) -> usize;
}

Trait for finding the minimum and maximum values in an array. For floats, NaNs are propagated - index of the first NaN is returned.

This trait is implemented for slices (or other array-like) of floats.

See the feature documentation for more information on the supported data types and array types.

Required Methods

fn nanargminmax(&self) -> (usize, usize)

Get the index of the minimum and maximum values in the array.

When dealing with floats, NaNs are propagated - index of the first NaN is returned.
Note that this differs from numpy, where the nanargmin and nanargmax functions ignore NaNs (which is the behavior of our argminmax function).

Returns

A tuple of the index of the minimum and maximum values in the array (min_index, max_index).

Caution

When multiple bit-representations for NaNs are used, no guarantee is made that the first NaN is returned.

fn nanargmin(&self) -> usize

Get the index of the minimum value in the array.

When dealing with floats, NaNs are propagated - index of the first NaN is returned. Note that this differs from numpy, where the nanargmin function ignores NaNs (which is the behavior of our argmin function).

Returns

The index of the minimum value in the array.

Caution

When multiple bit-representations for NaNs are used, no guarantee is made that the first NaN is returned.

fn nanargmax(&self) -> usize

Get the index of the maximum value in the array.

When dealing with floats, NaNs are propagated - index of the first NaN is returned. Note that this differs from numpy, where the nanargmax function ignores NaNs (which is the behavior of our argmax function).

Returns

The index of the maximum value in the array.

Caution

When multiple bit-representations for NaNs are used, no guarantee is made that the first NaN is returned.