argminmax::simd

Trait SIMDArgMinMax

pub trait SIMDArgMinMax<ScalarDType, SIMDVecDtype, SIMDMaskDtype, const LANE_SIZE: usize, SCALAR>: SIMDCore<ScalarDType, SIMDVecDtype, SIMDMaskDtype, LANE_SIZE>
where
    ScalarDType: Copy + PartialOrd + AsPrimitive<usize>,
    SIMDVecDtype: Copy,
    SIMDMaskDtype: Copy,
    SCALAR: ScalarArgMinMax<ScalarDType>,
{
    // Required methods
    unsafe fn argminmax(data: &[ScalarDType]) -> (usize, usize);
    unsafe fn argmin(data: &[ScalarDType]) -> usize;
    unsafe fn argmax(data: &[ScalarDType]) -> usize;
}

A trait providing the SIMD implementation of the argminmax operations.

Required Methods

unsafe fn argminmax(data: &[ScalarDType]) -> (usize, usize)

Get the index of the minimum and maximum values in the slice.

Arguments

data - the slice of data.

Returns

A tuple of the index of the minimum and maximum values in the slice (min_index, max_index).

Safety

This function is unsafe because unsafe SIMD operations are used.
See SIMD operations for more information:

x86 SIMD docs

x86_64 SIMD docs

arm SIMD docs

aarch64 SIMD docs

unsafe fn argmin(data: &[ScalarDType]) -> usize

Get the index of the minimum value in the slice.

Arguments

data - the slice of data.

Returns

The index of the minimum value in the slice.

Safety

This function is unsafe because unsafe SIMD operations are used.
See SIMD operations for more information:

x86 SIMD docs

x86_64 SIMD docs

arm SIMD docs

aarch64 SIMD docs

unsafe fn argmax(data: &[ScalarDType]) -> usize

Get the index of the maximum value in the slice.

Arguments

data - the slice of data.

Returns

The index of the maximum value in the slice.

Safety

This function is unsafe because unsafe SIMD operations are used.
See SIMD operations for more information:

x86 SIMD docs

x86_64 SIMD docs

arm SIMD docs

aarch64 SIMD docs

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.