Trait argminmax::ArgMinMax

pub trait ArgMinMax {
    // Required methods
    fn argminmax(&self) -> (usize, usize);
    fn argmin(&self) -> usize;
    fn argmax(&self) -> usize;
}

Trait for finding the minimum and maximum values in an array. For floats, NaNs are ignored.

This trait is implemented for slices (or other array-like) of integers and floats.

See the feature documentation for more information on the supported data types and array types.

Required Methods

fn argminmax(&self) -> (usize, usize)

Get the index of the minimum and maximum values in the array.

When dealing with floats, NaNs are ignored.
Note that this differs from numpy, where the argmin and argmax functions return the index of the first NaN (which is the behavior of our nanargminmax function).

Returns

A tuple of the index of the minimum and maximum values in the array (min_index, max_index).

Caution

When a float array contains only NaNs and / or infinities unexpected behavior may occur (in which case index 0 is returned for both).

fn argmin(&self) -> usize

Get the index of the minimum value in the array.

When dealing with floats, NaNs are ignored. Note that this differs from numpy, where the argmin function returns the index of the first NaN (which is the behavior of our nanargmin function).

Returns

The index of the minimum value in the array.

Caution

When a float array contains only NaNs and / or infinities unexpected behavior may occur (in which case index 0 is returned).

fn argmax(&self) -> usize

Get the index of the maximum value in the array.

When dealing with floats, NaNs are ignored. Note that this differs from numpy, where the argmax function returns the index of the first NaN (which is the behavior of our nanargmax function).

Returns

The index of the maximum value in the array.

Caution

When a float array contains only NaNs and / or infinities unexpected behavior may occur (in which case index 0 is returned).